Extrais la création de l’application express
...afin de pouvoir tester à l’aide de supertest
